A2/requirements.txt
fastapi==0.111.0
mysql-connector-python
pydantic==2.7.3
uvicorn==0.30.1
